+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org).
 
+## Unreleased
+
+### Added
+
+  - Oauth module.
+  - User, Client and Token models.
+  - Oauth management views.
+
 ## 0.3.0 - 2016-12-15
 
 ### Added